rag-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pypdf==4.2.0
python-multipart==0.0.9
numpy<2.0
